api를 가져와 데이터 정보 화면에 나타내기
