+++ src/index.js
@@ -1,9 +1,10 @@
 import ReactDOM from "react-dom/client";
-import App_copy from "./components/App_copy";
+import App from "./App";
+import Folder from "./page/folder/Folder";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
   <>
-    <App_copy />
+    <Folder />
   </>
 );
